# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  git_neolidy_top_neo_flowx_pkg_response.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  internal_interfaces_http_handler.approveRequest:
    properties:
      comment:
        type: string
    type: object
...
  internal_interfaces_http_handler.createWorkflowRequest:
    properties:
      definition:
        additionalProperties: {}
        type: object
      description:
        type: string
      name:
        maxLength: 200
        type: string
      type:
        maxLength: 50
        type: string
    required:
    - definition
    - name
    - type
    type: object
  internal_interfaces_http_handler.forwardRequest:
    properties:
      comment:
        type: string
      to_approver_id:
        type: string
    required:
    - to_approver_id
    type: object
...
  internal_interfaces_http_handler.rejectRequest:
    properties:
      comment:
        type: string
    required:
    - comment
    type: object
  internal_interfaces_http_handler.startApprovalRequest:
    properties:
      context:
        additionalProperties: {}
        type: object
      title:
        maxLength: 500
        type: string
      workflow_id:
        type: string
    required:
    - title
    - workflow_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: neolidy@outlook.com
    name: jiangfire
    url: https://github.com/jiangfire/flowx
  description: 企业级智能工具治理与流程编排平台 API 文档
  license:
    name: MIT
    url: https://github.com/jiangfire/flowx/blob/main/LICENSE
  termsOfService: https://github.com/jiangfire/flowx
  title: FlowX API
  version: "1.0"
paths:
...
  /approvals:
    get:
      consumes:
      - application/json
      description: 分页查询审批实例列表，支持按状态、工作流筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 实例状态
        in: query
        name: status
        type: string
      - description: 工作流ID
        in: query
        name: workflow_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 审批实例列表
      tags:
      - 审批工作流
    post:
      consumes:
      - application/json
      description: 基于工作流发起审批实例
      parameters:
      - description: 发起审批请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.startApprovalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 发起审批
      tags:
      - 审批工作流
  /approvals/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取审批实例详细信息
      parameters:
      - description: 审批实例ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 审批实例详情
      tags:
      - 审批工作流
  /approvals/{id}/approve:
    post:
      consumes:
      - application/json
      description: 审批通过指定审批实例
      parameters:
      - description: 审批实例ID
        in: path
        name: id
        required: true
        type: string
      - description: 审批通过请求参数
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.approveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 审批通过
      tags:
      - 审批工作流
  /approvals/{id}/cancel:
    post:
      consumes:
      - application/json
      description: 取消指定审批实例
      parameters:
      - description: 审批实例ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 取消审批
      tags:
      - 审批工作流
  /approvals/{id}/forward:
    post:
      consumes:
      - application/json
      description: 将审批实例转审给其他审批人
      parameters:
      - description: 审批实例ID
        in: path
        name: id
        required: true
        type: string
      - description: 转审请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.forwardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 转审
      tags:
      - 审批工作流
  /approvals/{id}/reject:
    post:
      consumes:
      - application/json
      description: 驳回指定审批实例
      parameters:
      - description: 审批实例ID
        in: path
        name: id
        required: true
        type: string
      - description: 审批驳回请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.rejectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 审批驳回
      tags:
      - 审批工作流
  /approvals/{id}/suggestion:
    get:
      consumes:
      - application/json
      description: 获取AI对指定审批实例的建议
      parameters:
      - description: 审批实例ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取AI审批建议
      tags:
      - 审批工作流
  /approvals/pending:
    get:
      consumes:
      - application/json
      description: 获取当前用户的待审批列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 我的待审批列表
      tags:
      - 审批工作流
...
  /workflows:
    get:
      consumes:
      - application/json
      description: 分页查询工作流列表，支持按类型、状态筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 工作流类型
        in: query
        name: type
        type: string
      - description: 工作流状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 工作流列表
      tags:
      - 审批工作流
    post:
      consumes:
      - application/json
      description: 创建新的审批工作流定义
      parameters:
      - description: 创建工作流请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.createWorkflowRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 创建工作流
      tags:
      - 审批工作流
  /workflows/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取工作流详细信息
      parameters:
      - description: 工作流ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 工作流详情
      tags:
      - 审批工作流
  /workflows/{id}/activate:
    post:
      consumes:
      - application/json
      description: 激活指定的工作流
      parameters:
      - description: 工作流ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 激活工作流
      tags:
      - 审批工作流
  /workflows/{id}/archive:
    post:
      consumes:
      - application/json
      description: 归档指定的工作流
      parameters:
      - description: 工作流ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 归档工作流
      tags:
      - 审批工作流
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer Token，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"